+++ model/provenancelinkdto.go
@@ -5,5 +5,5 @@ type ProvenanceLinkDTO struct {
 	TargetId     string `json:"targetId"`     // The target node id of the link.
 	FlowFileUuid string `json:"flowFileUuid"` // The flowfile uuid that traversed the link.
 	Timestamp    string `json:"timestamp"`    // The timestamp of the link (based on the destination).
-	Millis       int    `json:"millis"`       // The timestamp of this link in milliseconds.
+	Millis       int64  `json:"millis"`       // The timestamp of this link in milliseconds.
 }
