+++ model/queuesizedto.go
@@ -1,6 +1,6 @@
 package model
 
 type QueueSizeDTO struct {
-	ByteCount   int `json:"byteCount"`   // The size of objects in a queue.
-	ObjectCount int `json:"objectCount"` // The count of objects in a queue.
+	ByteCount   int64 `json:"byteCount"`   // The size of objects in a queue.
+	ObjectCount int32 `json:"objectCount"` // The count of objects in a queue.
 }
